renaming the variables and adding an input for the native edit context
