+++ src/vs/editor/browser/controller/editContext/native/nativeEditContextHandler.css
